Advanced org search.
* Functions to get the other dropdowns.
* Make dropdowns optional and add an 'Any' otherwise searches can be too
specific to return any results.
* Function to remove option items with value of 0, this allows you to
choose 'Any' and that option will be omitted from VC Connect API call.
* SECURITY: In the organisation list the template can decide whether to
display a record based on values for 'Expired' and
'InformationMadePublic'.  However, these values are not present when
making calls to the org details endpoint and therefore if you guessed an
org id you could potentially bypass this check and end up displaying
info that has not been made public.  To address this a new function that
checks if a record is displayable by calling the org list API endpoint
again is used in an if statement before showing any org details.